chore(ws): extract filter by labels component on workspace creation (#243)
