+++ workspaces/frontend/src/app/pages/Workspaces/Creation/image/WorkspaceCreationImageList.tsx
